Library

Documentation for SpaceIndices.jl.

SpaceIndices.SpaceIndexFile — Type

abstract type SpaceIndexFile

Abstract type for all structures that represent space files.

SpaceIndices.destroy_space_indices — Method

destroy_space_indices()

Destroy the object of all space indices that were initialized.

SpaceIndices.fetch_space_file — Method

fetch_space_file(::Type{T}; kwargs...) where T<:SpaceIndexFile -> String

Fetch the space file related to the space index T. This function returns the space index file path.

Keywords

force_download::Bool: If true, the space file will be downloaded regardless of its timestamp. (Default = false)

SpaceIndices.get_expiry_period — Method

get_expiry_period(::Type{T}) where T<:SpaceIndexFile -> DatePeriod

Return the expiry period for the space index file T. The remote file will always be downloaded again if a time larger than this period has passed after the last download.

If this function is not defined for T, the default expiry period of 7 days is used.

SpaceIndices.get_filename — Function

get_filename(::Type{T}) where T<:SpaceIndexFile -> String

Return the filename for the space index file T.

SpaceIndices.get_space_index — Function

get_space_index(::Val{:index}, jd::Number; kwargs...) -> Number
get_space_index(::Val{:index}, instant::DateTime; kwargs...) -> Number

Get the space index for the Julian day jd or instant, which can be an object of type DateTime. kwargs... can be used to pass additional configuration for the space index.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:DTC}, jd_utc::Number) -> Float64

Get the exospheric temperature variation [K] caused by the Dst index.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:F10adj}, jd_utc::Number) -> Float64

Get the adjusted F10.7 index (10.7-cm solar flux) [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:F10obs}, jd_utc::Number) -> Float64

Get the observed F10.7 index (10.7-cm solar flux) [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:M10}, jd_utc::Number) -> Float64

Get the MG2 index scaled to F10.7 [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:M81a}, jd_utc::Number) -> Float64

Get the 81-day averaged MG2 index scaled to F10.7 [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:S10}, jd_utc::Number) -> Float64

Get the EUV index (26-34 nm) scaled to F10.7 [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:S81a}, jd_utc::Number) -> Float64

Get the 81-day averaged EUV index (26-34 nm) scaled to F10.7 [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:Y10}, jd_utc::Number) -> Float64

Get the solar X-ray & Lya index scaled to F10.7 [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_space_index — Method

get_space_index(::Val{:Y81a}, jd_utc::Number) -> Float64

Get the 81-day averaged solar X-ray & Lya index scaled to F10.7 [10⁻²² W / (M² ⋅ Hz)] for the Julian day jd_utc.

SpaceIndices.get_url — Function

get_url(::Type{T}) where T<:SpaceIndexFile -> String

Return the URL to obtain the space index file T.

SpaceIndices.init_space_index — Method

init_space_index(::Type{T}; kwargs...) where T<:SpaceIndexFile

Initialize the space index T.

This function will download the remote file associated to the space index T if it does not exist or if the redownload period has been passed. Aftward, it will parse the file and populate the object handler to be accesses by the function get_space_index.

Keywords

force_download::Bool: If true, the space file will be downloaded regardless of its timestamp. (Default = false)

SpaceIndices.init_space_indices — Method

init_space_indices(; blocklist::Vector = []) -> Nothing

Initialize all the registered space indices.

This function will download the remote files associated to the space indices if they do not exist or if the redownload period has been passed. Aftward, it will parse the files and populate the object handlers to be accesses by the function get_space_index.

SpaceIndices.parse_space_file — Function

parse_space_file(::Type{T}, filepath::String) where T<:SpaceIndexFile -> T

Parse the space index file T using the file in filepath. It must return an object of type T with the parsed data.

SpaceIndices.@check_timespan — Macro

@check_timespan(itp, jd)

Check if the interpolation itp contains the date jd within its nodes. If not, it thrown an ArgumentError exception.

SpaceIndices.@data_handler — Macro

@data_handler(T)

Get the variable with the optional data handler for space index file structure T.

SpaceIndices.@object — Macro

@object(T)

Get the data handler for the space index file structure T.

SpaceIndices.@register — Macro

@register(T)

Register the the space index file structure T. This macro push the data into the global vector of space files and also creates the optinal data handler for the processed structure.